+++ languagetool-language-modules/ca/src/main/java/org/languagetool/rules/ca/AdjustVerbSuggestionsFilter.java
@@ -186,7 +186,7 @@ public class AdjustVerbSuggestionsFilter extends RuleFilter {
           replacement = doAddPronounReflexive("", "hi " + verbStr, firstVerbPersonaNumber, isPronounsAfter);
           break;
         case "addPronounReflexiveLes":
-          replacement = doAddPronounReflexive("les", verbStr, firstVerbPersonaNumber, isPronounsAfter);
+          replacement = doAddPronounReflexive(transform(pronounsStr.toLowerCase(), PronounPosition.NORMALIZED) + " les", verbStr, firstVerbPersonaNumber, isPronounsAfter);
           break;
         case "addPronounDative":
           String dativePronoun = getDativePronoun(firstVerbPersonaNumber);
@@ -265,7 +265,7 @@ public class AdjustVerbSuggestionsFilter extends RuleFilter {
       }
     }
     // extend the ending position if there are pronouns after the verb
-    int endingPos = Math.max(match.getToPos(), tokens[verbSynthesizer.getLastIndex()].getStartPos());
+    int endingPos = Math.max(match.getToPos(), tokens[verbSynthesizer.getLastIndex()].getEndPos());
     RuleMatch ruleMatch = new RuleMatch(match.getRule(), match.getSentence(), tokens[posStartUnderline].getStartPos(),
       endingPos, match.getMessage(), match.getShortMessage());
     ruleMatch.setType(match.getType());