+++ languagetool-language-modules/ca/src/test/java/org/languagetool/JLanguageToolTest.java
@@ -267,6 +267,7 @@ public class JLanguageToolTest {
 
   @Test
   public void testCatalanLongSentenceRule() throws IOException {
+    tool.enableRule("CA_SPLIT_LONG_SENTENCE");
     List<RuleMatch> matches = tool.check(
       "En una tarda grisa que avançava sense pressa sobre els carrers estrets de la ciutat, mentre els comerços " +
         "abaixaven persianes i el soroll del trànsit es diluïa en un murmuri constant, un home caminava pensant en " +
@@ -275,7 +276,7 @@ public class JLanguageToolTest {
         "ordenar les idees, assumir els errors, fer servir l’experiència acumulada com a criteri i continuar avançant" +
         " amb una determinació menys impulsiva però més sòlida."
       , JLanguageTool.Level.PICKY);
-    assertEquals(matches.get(0).getSuggestedReplacements().toString(), "[desfigurant. I]");
+    assertEquals("[desfigurant. I]", matches.get(0).getSuggestedReplacements().toString());
   }
 
   @Test