+++ languagetool-language-modules/uk/src/test/java/org/languagetool/rules/uk/TokenAgreementPrepNounRuleTest.java
@@ -122,6 +122,7 @@ public class TokenAgreementPrepNounRuleTest extends AbstractRuleTest {
     assertEmptyMatch("станом на зараз виконавча влада");
 
     assertEmptyMatch("в тисяча шістсот якомусь році");
+    assertEmptyMatch("у ти́сяча сімсо́т вісімдеся́т дев’я́тому");
     assertEmptyMatch("на Піп Іван");
 
 //    assertEmptyMatch("Імена від Андрій до Юрій");  // називний між від і до рідко зустрічається але такий виняток ховає багато помилок